Change from the button format to google
Changes agenda format from the button layout to the google calendar.  Once the email for the conference is created, this calendar can be imported to the google account of the conference email. Once there everyone with access to the google account will have the ability to add and change events. This version is set so that events are displayed in Agenda format. We are also able to share the calendar with personal emails so that everyone can access it outside of the conference email if necessary.